+++ CHANGELOG.md
@@ -2,7 +2,29 @@
 
 ## [Unreleased]
 
-## [2.0.1] - 2017-08-07
+## [2.1.0-pre] - 2017-10-08
+
+### Added
+
+- Add support for `.NET Standard 2.0`
+- Add `OnError` method to `NetServer`
+- Add array reader on `NetPacketBase` (`T[] Read<T>(amount)`)
+
+### Changed
+
+- Change sending process on `NetServer`, now uses a sending queue
+
+### Fix
+
+- Call the `OnClientDisconnected` when the `DisconnectClient(Guid id)` method is called and the connection is disposed.
+
+### Removed
+
+- Buffer Manager
+
+## [Released]
+
+## [2.0.1] - 2017-08-06
 
 ### Added
 
@@ -11,8 +33,6 @@
 ### Changed
 
 - Changed `ConcurrentBag` to `ConcurrentDictionary` for client handling.
-
-## [Released]
 
 ## [2.0.0] - 2017-08-01